requirements.txt
beautifulsoup4==4.6.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
django-modelcluster==4.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.10.3
draftjs-exporter==2.1.7
gunicorn==20.0.0
html5lib==1.0.1
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
#pkg-resources==0.0.0
psycopg2==2.8.4

python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.6
wagtail==2.6.2
webencodings==0.5.1
Willow==1.1
